Refactor GHA verify logic
