+++ pkg/internal/appco/verifier.go
@@ -26,8 +26,9 @@ var (
 )
 
 type Verifier struct {
-	HashAlgorithm crypto.Hash
 	internal.UpstreamVerifier
+
+	HashAlgorithm crypto.Hash
 }
 
 func (v *Verifier) Matches(image string) bool {
