+++ pkg/internal/obs/verifier.go
@@ -37,8 +37,9 @@ var (
 )
 
 type Verifier struct {
-	HashAlgorithm crypto.Hash
 	internal.UpstreamVerifier
+
+	HashAlgorithm crypto.Hash
 }
 
 func (v *Verifier) Matches(image string) bool {
